requirements-cxfreeze38.txt
# Pure Python 3.8 + venv migration toolchain for cx_Freeze packaging
cx-Freeze==7.2.5

# Application runtime dependencies pinned to the currently verified desktop stack
PyQt5==5.15.9
PyQt-Fluent-Widgets==1.10.5
PyQt5-Frameless-Window==0.7.5
paramiko==2.9.2
matplotlib==3.3.4
pyte==0.8.2
python-docx==1.1.2
pywin32==311
wcwidth==0.2.14
